Rename and fix PregReplaceWithEvalModifier to handle different delimiters
